Amend neuronym from kb8 to k8s
It's generally accepted that the neuronym for Kubernetes is k8s (as between
'k' and 's' there are eight characters). To adhere to this, because I'm a
bit of a neuronym pedant, this proposes moving the kb8/ directory to k8s/
and updating the README.

I couldn't find anywhere else that kb8/ was referenced and therefore needed
changing, but I didn't unpack any binaries to see if anything was set
anywhere else.